combine PluginRedirect and TemplateUrlHelper into one RedirectHelper
this allows for live/dynamic template links which could become outdated before
we also update all template links to use the redirect and plugin links to include updated params
