+++ includes/Data/Plans/BlogPlan.php
@@ -3,8 +3,7 @@
 namespace NewfoldLabs\WP\Module\NextSteps\Data\Plans;
 
 use NewfoldLabs\WP\Module\NextSteps\DTOs\Plan;
-use NewfoldLabs\WP\Module\NextSteps\PluginRedirect;
-use NewfoldLabs\WP\Module\NextSteps\TemplateUrlHelper;
+use NewfoldLabs\WP\Module\NextSteps\RedirectHelper;
 
 /**
  * BlogPlan - Defines the structured plan for blog/personal website setup
@@ -80,7 +79,7 @@ class BlogPlan {
 									array(
 										'id'       => 'blog_upload_logo',
 										'title'    => __( 'Upload Logo', 'wp-module-next-steps' ),
-										'href'     => TemplateUrlHelper::get_url_to_active_template_editor( 'header' ),
+										'href'     => '{siteUrl}/wp-admin/admin.php?page=redirect-check&type=template&area=header&template=index',
 										'status'   => 'new',
 										'priority' => 1,
 										'source'   => 'wp-module-next-steps',
@@ -96,7 +95,7 @@ class BlogPlan {
 									array(
 										'id'       => 'blog_customize_header',
 										'title'    => __( 'Customize Header', 'wp-module-next-steps' ),
-										'href'     => TemplateUrlHelper::get_url_to_active_template_editor( 'header' ),
+										'href'     => '{siteUrl}/wp-admin/admin.php?page=redirect-check&type=template&area=header&template=index',
 										'status'   => 'new',
 										'priority' => 3,
 										'source'   => 'wp-module-next-steps',
@@ -104,7 +103,7 @@ class BlogPlan {
 									array(
 										'id'       => 'blog_customize_footer',
 										'title'    => __( 'Customize Footer', 'wp-module-next-steps' ),
-										'href'     => TemplateUrlHelper::get_url_to_active_template_editor( 'footer' ),
+										'href'     => '{siteUrl}/wp-admin/admin.php?page=redirect-check&type=template&area=footer&template=index',
 										'status'   => 'new',
 										'priority' => 3,
 										'source'   => 'wp-module-next-steps',
@@ -223,7 +222,7 @@ class BlogPlan {
 									array(
 										'id'       => 'blog_customize_notification_emails',
 										'title'    => __( 'Customize Notification Emails', 'wp-module-next-steps' ),
-										'href'     => '{siteUrl}/wp-admin/admin.php?page=redirect-check&p=email-templates',
+										'href'     => '{siteUrl}/wp-admin/admin.php?page=redirect-check&type=plugin&p=email-templates',
 										'status'   => 'new',
 										'priority' => 2,
 										'source'   => 'wp-module-next-steps',
@@ -231,7 +230,7 @@ class BlogPlan {
 									array(
 										'id'       => 'blog_connect_jetpack_stats',
 										'title'    => __( 'Connect Jetpack Stats (or Google Analytics 4)', 'wp-module-next-steps' ),
-										'href'     => '{siteUrl}/wp-admin/admin.php?page=redirect-check&p=jetpack&r=' . rawurlencode( 'admin.php?page=stats' ),
+										'href'     => '{siteUrl}/wp-admin/admin.php?page=redirect-check&type=plugin&p=jetpack&r=' . rawurlencode( 'admin.php?page=stats' ),
 										'status'   => 'new',
 										'priority' => 3,
 										'source'   => 'wp-module-next-steps',
@@ -245,7 +244,7 @@ class BlogPlan {
 									array(
 										'id'       => 'blog_connect_facebook',
 										'title'    => __( 'Connect Facebook Page Auto-Sharing', 'wp-module-next-steps' ),
-										'href'     => '{siteUrl}/wp-admin/admin.php?page=redirect-check&p=jetpack&r=' . rawurlencode( 'admin.php?page=jetpack-social' ),
+										'href'     => '{siteUrl}/wp-admin/admin.php?page=redirect-check&type=plugin&p=jetpack&r=' . rawurlencode( 'admin.php?page=jetpack-social' ),
 										'status'   => 'new',
 										'priority' => 1,
 										'source'   => 'wp-module-next-steps',
@@ -253,7 +252,7 @@ class BlogPlan {
 									array(
 										'id'       => 'blog_add_social_sharing',
 										'title'    => __( 'Add Social-Sharing Buttons', 'wp-module-next-steps' ),
-										'href'     => '{siteUrl}/wp-admin/admin.php?page=redirect-check&p=jetpack&r=' . rawurlencode( 'site-editor.php?p=%2Fstyles&section=%2Fblocks%2Fjetpack%252Fsharing-buttons' ),
+										'href'     => '{siteUrl}/wp-admin/admin.php?page=redirect-check&type=plugin&p=jetpack&r=' . rawurlencode( 'site-editor.php?p=%2Fstyles&section=%2Fblocks%2Fjetpack%252Fsharing-buttons' ),
 										'status'   => 'new',
 										'priority' => 2,
 										'source'   => 'wp-module-next-steps',
@@ -275,7 +274,7 @@ class BlogPlan {
 									array(
 										'id'       => 'blog_optimize_seo',
 										'title'    => __( 'Optimize On-Page SEO', 'wp-module-next-steps' ),
-										'href'     => '{siteUrl}/wp-admin/admin.php?page=redirect-check&p=yoast-seo',
+										'href'     => '{siteUrl}/wp-admin/admin.php?page=redirect-check&type=plugin&p=yoast-seo',
 										'status'   => 'new',
 										'priority' => 1,
 										'source'   => 'wp-module-next-steps',
@@ -311,7 +310,7 @@ class BlogPlan {
 									array(
 										'id'       => 'blog_enable_comments',
 										'title'    => __( 'Enable & Style Comments Section', 'wp-module-next-steps' ),
-										'href'     => '{siteUrl}/wp-admin/admin.php?page=redirect-check&p=akismet',
+										'href'     => '{siteUrl}/wp-admin/admin.php?page=redirect-check&type=plugin&p=akismet',
 										'status'   => 'new',
 										'priority' => 1,
 										'source'   => 'wp-module-next-steps',
@@ -403,7 +402,7 @@ class BlogPlan {
 									array(
 										'id'       => 'blog_install_yoast_premium',
 										'title'    => __( 'Install Yoast Premium for Advanced Schemas', 'wp-module-next-steps' ),
-										'href'     => '{siteUrl}/wp-admin/admin.php?page=redirect-check&p=yoast-seo',
+										'href'     => '{siteUrl}/wp-admin/admin.php?page=redirect-check&type=plugin&p=yoast-seo',
 										'status'   => 'new',
 										'priority' => 3,
 										'source'   => 'wp-module-next-steps',
@@ -417,7 +416,7 @@ class BlogPlan {
 									array(
 										'id'       => 'blog_speed_up_site',
 										'title'    => __( 'Speed-up Site with Jetpack Boost', 'wp-module-next-steps' ),
-										'href'     => '{siteUrl}/wp-admin/admin.php?page=redirect-check&p=jetpack&r=' . rawurlencode( 'my-jetpack#/add-boost' ),
+										'href'     => '{siteUrl}/wp-admin/admin.php?page=redirect-check&type=plugin&p=jetpack&r=' . rawurlencode( 'my-jetpack#/add-boost' ),
 										'status'   => 'new',
 										'priority' => 1,
 										'source'   => 'wp-module-next-steps',
@@ -425,7 +424,7 @@ class BlogPlan {
 									array(
 										'id'       => 'blog_enable_auto_backups',
 										'title'    => __( 'Enable Automatic Backups & Update Alerts', 'wp-module-next-steps' ),
-										'href'     => '{siteUrl}/wp-admin/admin.php?page=redirect-check&p=jetpack&r=' . rawurlencode( 'admin.php?page=jetpack-backup' ),
+										'href'     => '{siteUrl}/wp-admin/admin.php?page=redirect-check&type=plugin&p=jetpack&r=' . rawurlencode( 'admin.php?page=jetpack-backup' ),
 										'status'   => 'new',
 										'priority' => 2,
 										'source'   => 'wp-module-next-steps',
@@ -447,7 +446,7 @@ class BlogPlan {
 									array(
 										'id'       => 'blog_monitor_traffic',
 										'title'    => __( 'Monitor Traffic & Engagement', 'wp-module-next-steps' ),
-										'href'     => '{siteUrl}/wp-admin/admin.php?page=redirect-check&p=jetpack&r=' . rawurlencode( 'admin.php?page=stats' ),
+										'href'     => '{siteUrl}/wp-admin/admin.php?page=redirect-check&type=plugin&p=jetpack&r=' . rawurlencode( 'admin.php?page=stats' ),
 										'status'   => 'new',
 										'priority' => 1,
 										'source'   => 'wp-module-next-steps',
