+++ includes/Data/Plans/StorePlan.php
@@ -3,8 +3,7 @@
 namespace NewfoldLabs\WP\Module\NextSteps\Data\Plans;
 
 use NewfoldLabs\WP\Module\NextSteps\DTOs\Plan;
-use NewfoldLabs\WP\Module\NextSteps\PluginRedirect;
-use NewfoldLabs\WP\Module\NextSteps\TemplateUrlHelper;
+use NewfoldLabs\WP\Module\NextSteps\RedirectHelper;
 
 /**
  * StorePlan - Defines the structured plan for ecommerce store setup
@@ -52,7 +51,7 @@ class StorePlan {
 										'id'              => 'store_upload_logo',
 										'title'           => __( 'Upload Logo', 'wp-module-next-steps' ),
 										'description'     => '',
-										'href'            => TemplateUrlHelper::get_url_to_active_template_editor( 'header' ),
+										'href'            => '{siteUrl}/wp-admin/admin.php?page=redirect-check&type=template&area=header&template=index',
 										'status'          => 'new',
 										'priority'        => 1,
 										'source'          => 'wp-module-next-steps',
@@ -72,7 +71,7 @@ class StorePlan {
 										'id'              => 'store_customize_header',
 										'title'           => __( 'Customize Header', 'wp-module-next-steps' ),
 										'description'     => '',
-										'href'            => TemplateUrlHelper::get_url_to_active_template_editor( 'header' ),
+										'href'            => '{siteUrl}/wp-admin/admin.php?page=redirect-check&type=template&area=header&template=index',
 										'status'          => 'new',
 										'priority'        => 3,
 										'source'          => 'wp-module-next-steps',
@@ -82,7 +81,7 @@ class StorePlan {
 										'id'              => 'store_customize_footer',
 										'title'           => __( 'Customize Footer', 'wp-module-next-steps' ),
 										'description'     => '',
-										'href'            => TemplateUrlHelper::get_url_to_active_template_editor( 'footer' ),
+										'href'            => '{siteUrl}/wp-admin/admin.php?page=redirect-check&type=template&area=footer&template=index',
 										'status'          => 'new',
 										'priority'        => 4,
 										'source'          => 'wp-module-next-steps',
@@ -92,7 +91,7 @@ class StorePlan {
 										'id'              => 'store_customize_homepage',
 										'title'           => __( 'Customize Homepage', 'wp-module-next-steps' ),
 										'description'     => '',
-										'href'            => TemplateUrlHelper::get_url_to_home_template_editor(),
+										'href'            => '{siteUrl}/wp-admin/admin.php?page=redirect-check&type=template&template=home',
 										'status'          => 'new',
 										'priority'        => 5,
 										'source'          => 'wp-module-next-steps',
@@ -138,7 +137,7 @@ class StorePlan {
 										'id'              => 'store_setup_payments',
 										'title'           => __( 'Set up Payments', 'wp-module-next-steps' ),
 										'description'     => '',
-										'href'            => '{siteUrl}/wp-admin/admin.php?page=redirect-check&p=woocommerce&r=' . rawurlencode( 'admin.php?page=wc-settings&tab=checkout' ),
+										'href'            => '{siteUrl}/wp-admin/admin.php?page=redirect-check&type=plugin&p=woocommerce&r=' . rawurlencode( 'admin.php?page=wc-settings&tab=checkout' ),
 										'status'          => 'new',
 										'priority'        => 1,
 										'source'          => 'wp-module-next-steps',
@@ -159,7 +158,7 @@ class StorePlan {
 										'id'              => 'store_customize_shop_page',
 										'title'           => __( 'Customize the shop page', 'wp-module-next-steps' ),
 										'description'     => '',
-										'href'            => '{siteUrl}/wp-admin/admin.php?page=redirect-check&p=woocommerce&r=' . rawurlencode( 'admin.php?page=wc-settings&tab=products' ),
+										'href'            => '{siteUrl}/wp-admin/admin.php?page=redirect-check&type=plugin&p=woocommerce&r=' . rawurlencode( 'admin.php?page=wc-settings&tab=products' ),
 										'status'          => 'new',
 										'priority'        => 1,
 										'source'          => 'wp-module-next-steps',
@@ -169,7 +168,7 @@ class StorePlan {
 										'id'              => 'store_customize_cart_page',
 										'title'           => __( 'Customize the cart page', 'wp-module-next-steps' ),
 										'description'     => '',
-										'href'            => '{siteUrl}/wp-admin/admin.php?page=redirect-check&p=woocommerce',
+										'href'            => '{siteUrl}/wp-admin/admin.php?page=redirect-check&type=plugin&p=woocommerce',
 										'status'          => 'new',
 										'priority'        => 2,
 										'source'          => 'wp-module-next-steps',
@@ -179,7 +178,7 @@ class StorePlan {
 										'id'              => 'store_customize_checkout_page',
 										'title'           => __( 'Customize the checkout flow', 'wp-module-next-steps' ),
 										'description'     => '',
-										'href'            => '{siteUrl}/wp-admin/admin.php?page=redirect-check&p=woocommerce',
+										'href'            => '{siteUrl}/wp-admin/admin.php?page=redirect-check&type=plugin&p=woocommerce',
 										'status'          => 'new',
 										'priority'        => 3,
 										'source'          => 'wp-module-next-steps',
@@ -200,7 +199,7 @@ class StorePlan {
 										'id'              => 'store_marketing_welcome_popup',
 										'title'           => __( 'Configure a welcome discount popup', 'wp-module-next-steps' ),
 										'description'     => '',
-										'href'            => '{siteUrl}/wp-admin/admin.php?page=redirect-check&p=wondercart',
+										'href'            => '{siteUrl}/wp-admin/admin.php?page=redirect-check&type=plugin&p=wondercart',
 										'status'          => 'new',
 										'priority'        => 1,
 										'source'          => 'wp-module-next-steps',
@@ -210,7 +209,7 @@ class StorePlan {
 										'id'              => 'store_create_gift_card',
 										'title'           => __( 'Create a gift card to sell in your shop', 'wp-module-next-steps' ),
 										'description'     => '',
-										'href'            => '{siteUrl}/wp-admin/admin.php?page=redirect-check&p=gift-cards',
+										'href'            => '{siteUrl}/wp-admin/admin.php?page=redirect-check&type=plugin&p=gift-cards',
 										'status'          => 'new',
 										'priority'        => 2,
 										'source'          => 'wp-module-next-steps',
@@ -220,7 +219,7 @@ class StorePlan {
 										'id'              => 'store_enable_abandoned_cart',
 										'title'           => __( 'Enable abandoned cart emails', 'wp-module-next-steps' ),
 										'description'     => '',
-										'href'            => '{siteUrl}/wp-admin/admin.php?page=redirect-check&p=email-templates',
+										'href'            => '{siteUrl}/wp-admin/admin.php?page=redirect-check&type=plugin&p=email-templates',
 										'status'          => 'new',
 										'priority'        => 3,
 										'source'          => 'wp-module-next-steps',
@@ -230,7 +229,7 @@ class StorePlan {
 										'id'              => 'store_customize_emails',
 										'title'           => __( 'Customize your store emails', 'wp-module-next-steps' ),
 										'description'     => '',
-										'href'            => '{siteUrl}/wp-admin/admin.php?page=redirect-check&p=email-templates',
+										'href'            => '{siteUrl}/wp-admin/admin.php?page=redirect-check&type=plugin&p=email-templates',
 										'status'          => 'new',
 										'priority'        => 4,
 										'source'          => 'wp-module-next-steps',
@@ -250,7 +249,7 @@ class StorePlan {
 										'id'              => 'store_improve_performance',
 										'title'           => __( 'Improve the performance and speed of your shop', 'wp-module-next-steps' ),
 										'description'     => '',
-										'href'            => '{siteUrl}/wp-admin/admin.php?page=redirect-check&p=jetpack',
+										'href'            => '{siteUrl}/wp-admin/admin.php?page=redirect-check&type=plugin&p=jetpack',
 										'status'          => 'new',
 										'priority'        => 1,
 										'source'          => 'wp-module-next-steps',
@@ -270,7 +269,7 @@ class StorePlan {
 										'id'              => 'store_collect_reviews_task',
 										'title'           => __( 'Collect and Show Reviews for Your Products', 'wp-module-next-steps' ),
 										'description'     => '',
-										'href'            => '{siteUrl}/wp-admin/admin.php?page=redirect-check&p=advanced-reviews',
+										'href'            => '{siteUrl}/wp-admin/admin.php?page=redirect-check&type=plugin&p=advanced-reviews',
 										'status'          => 'new',
 										'priority'        => 1,
 										'source'          => 'wp-module-next-steps',
@@ -290,7 +289,7 @@ class StorePlan {
 										'id'              => 'store_launch_affiliate',
 										'title'           => __( 'Launch an affiliate program', 'wp-module-next-steps' ),
 										'description'     => '',
-										'href'            => '{siteUrl}/wp-admin/admin.php?page=redirect-check&p=affiliates',
+										'href'            => '{siteUrl}/wp-admin/admin.php?page=redirect-check&type=plugin&p=affiliates',
 										'status'          => 'new',
 										'priority'        => 1,
 										'source'          => 'wp-module-next-steps',
@@ -310,7 +309,7 @@ class StorePlan {
 										'id'              => 'store_setup_yoast_premium',
 										'title'           => __( 'Setup Yoast Premium to drive traffic to your store', 'wp-module-next-steps' ),
 										'description'     => '',
-										'href'            => '{siteUrl}/wp-admin/admin.php?page=redirect-check&p=yoast-seo',
+										'href'            => '{siteUrl}/wp-admin/admin.php?page=redirect-check&type=plugin&p=yoast-seo',
 										'status'          => 'new',
 										'priority'        => 1,
 										'source'          => 'wp-module-next-steps',
